move databasetest.MockDBTime() & databasetest.RestoreDBTime() back into the testhelpers package (remove the databasetest package)
