+++ testhelpers/test_context.go
@@ -19,7 +19,6 @@ import (
 	"github.com/thingful/httpmock"
 
 	"github.com/France-ioi/AlgoreaBackend/app"
-	"github.com/France-ioi/AlgoreaBackend/app/databasetest"
 	log "github.com/France-ioi/AlgoreaBackend/app/logging"
 	"github.com/France-ioi/AlgoreaBackend/app/loggingtest"
 )
@@ -104,7 +103,7 @@ func (ctx *TestContext) tearDownApp() {
 }
 
 func (ctx *TestContext) ScenarioTeardown(interface{}, error) { // nolint
-	databasetest.RestoreDBTime()
+	RestoreDBTime()
 	monkey.UnpatchAll()
 	ctx.logsRestoreFunc()
